watermeter

+++ orchid/src/main/java/com/ih2ome/hardware_service/service/model/narcissus/SmartGateway.java
@@ -26,8 +26,8 @@ public class SmartGateway implements Serializable{
   private String installMobile;
   private String brand;
   private String operator;
-  private Double installStatus;
-  private Double onoffStatus;
+  private int installStatus;
+  private int onoffStatus;
   private String remark;
   private Long houseCatalog;
   private Long apartmentId;
