+++ orchid/src/main/java/com/ih2ome/hardware_service/service/model/narcissus/SmartWatermeter.java
@@ -8,7 +8,7 @@ import java.util.Date;
 @Data
 public class SmartWatermeter implements Serializable{
 
-  private Long smartWatermeterId;
+  private int smartWatermeterId;
   private Date createdAt;
   private Long createdBy;
   private Date updatedAt;
@@ -19,11 +19,10 @@ public class SmartWatermeter implements Serializable{
   private Long floorId;
   private Long houseId;
   private Long roomId;
-  private Long houseCatalog;
-  //TODO:double float
-  private Double meterType;
+  private int houseCatalog;
+  private int meterType;
   private String uuid;
-  private Double onoffStatus;
+  private int onoffStatus;
   private Long price;
   private Long meterAmount;
   private Long lastAmount;
